Move update sparta service to background job
